+++ CHANGELOG.md
@@ -13,6 +13,7 @@ expected to be especially active and follows a simplified Semantic Versioning:
 The changelog format is inspired by [keep-a-changelog](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased]
+- "sendto" now looks for a config file in home directory
 
 ## [1.0] - 2016-04-25